app/requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==7.0.0  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.4.2  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==4.1.5  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.52.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.2.0  # https://github.com/niwinz/django-redis
django-active-link==0.1.8 # https://django-active-link.readthedocs.io/en/latest/readme.html#quick-start

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Swagger API
drf-yasg==1.21.4

django-oauth-toolkit==2.2.0

# enable cors
django-cors-headers

# AWS S3 Integration
boto3==1.26.47
botocore==1.29.47
django-storages==1.13.2

celery[sqs]==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.4.0

# current user logging
django-crum==0.7.9

# Tables
django-tables2==2.5.1  # https://django-tables2.readthedocs.io/en/latest/
django-filter==22.1  # https://django-filter.readthedocs.io/en/stable/
django-sass==1.1.0

# Email Notifications
premailer==3.10.0